move functions that directly access the database into their own package
